Introduce weight util for shared methods
